hopefully this tasking system will be batter...

+++ simeng_task/src/flags.rs
@@ -1,13 +1,13 @@
-pub const SCHEDULED: u32 = 1 << 0;
-pub const RUNNING: u32 = 1 << 1;
-pub const COMPLETED: u32 = 1 << 2;
-pub const CLOSED: u32 = 1 << 3;
+pub const SCHEDULED: u16 = 1 << 0;
+pub const RUNNING: u16 = 1 << 1;
+pub const COMPLETED: u16 = 1 << 2;
+pub const CLOSED: u16 = 1 << 3;
 //pub const CANCELLED: u32 = 1 << 4;
-pub const TAKEN: u32 = 1 << 5;
-pub const HANDLE_HERE: u32 = 1 << 6;
-pub const WAKER_IN_HERE: u32 = 1 << 7;
-pub const WAKER_REGISTERING: u32 = 1 << 8;
-pub const WAKER_NOTIFYING: u32 = 1 << 9;
+pub const TAKEN: u16 = 1 << 5;
+pub const HANDLE_HERE: u16 = 1 << 6;
+pub const WAKER_IN_HERE: u16 = 1 << 7;
+pub const WAKER_REGISTERING: u16 = 1 << 8;
+pub const WAKER_NOTIFYING: u16 = 1 << 9;
 
 impl crate::state::State {
     pub const fn is_completed(&self) -> bool {
